+++ spec/functional/spreadsheets/spreadsheets_with_macros_to_ooxml_format_spec.rb
@@ -3,8 +3,8 @@
 require './spec/spec_helper'
 require 'nokogiri'
 
-palladium = PalladiumHelper.new DocumentServerHelper.get_version, 'Spreadsheets With Macros to Ooxml'
-result_sets = palladium.get_result_sets StaticData::POSITIVE_STATUSES
+report = TestReport.new(DocumentServerHelper.get_version, 'Spreadsheets With Macros to Ooxml')
+completed_tests = report.completed_tests(StaticData::POSITIVE_STATUSES)
 files = StaticData::SPREADSHEETS['spreadsheets_with_macros']
 
 describe 'Convert spreadsheets with macros to ooxml format by convert service' do
@@ -17,7 +17,7 @@ describe 'Convert spreadsheets with macros to ooxml format by convert service' d
     input_format = File.extname(s3_file_path).delete('.').to_s
 
     test_name = "#{input_format} with macros to ooxml"
-    next if result_sets.include?(test_name)
+    next if completed_tests.include?(test_name)
 
     it test_name do
       pending('https://bugzilla.onlyoffice.com/show_bug.cgi?id=61652') if input_format == 'ods'
@@ -35,6 +35,6 @@ describe 'Convert spreadsheets with macros to ooxml format by convert service' d
 
   after do |example|
     FileUtils.rm_rf(@tmp_dir, secure: true)
-    palladium.add_result_and_log(example)
+    report.add_result_and_log(example)
   end
 end